管理员用户管理功能完善

+++ src/main/java/com/czvisual/controller/UserController.java
@@ -1,5 +1,6 @@
 package com.czvisual.controller;
 
+import com.alibaba.fastjson.JSONObject;
 import com.czvisual.config.UserCredentialsMatcher;
 import com.czvisual.entity.User;
 import com.czvisual.service.SessionService;
@@ -18,6 +19,10 @@ import org.springframework.web.bind.annotation.ResponseBody;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 import static org.springframework.web.bind.annotation.RequestMethod.POST;
 
@@ -27,8 +32,9 @@ public class UserController {
     private UserService userService;
 
     @Autowired
+    private SessionService sessionService;
+    @Autowired
     HttpSession session;
-
 
     @RequestMapping("")
     public String defaultMapping() {
@@ -77,9 +83,9 @@ public class UserController {
     }
 
 
-    @RequestMapping("register")
+    @RequestMapping({"register"})
     @ResponseBody
-    public Object addUser(User user) {
+    public Object register(User user) {
         int i1 = userService.checkUser(user);
         if (i1 == 1) {
             return "当前登陆名已存在";
@@ -102,6 +108,49 @@ public class UserController {
         }
     }
 
+    @RequestMapping({"/user/addUser"})
+    @ResponseBody
+    public Object addUser(User user) {
+        JSONObject jo = new JSONObject();
+        int i1 = userService.checkUser(user);
+        if (i1 == 1) {
+            jo.put("code", 1);
+            jo.put("message", "当前登陆名已存在");
+            return jo;
+        } else {
+            //加盐
+            String salt = UserCredentialsMatcher.generateSalt(6);
+            //MD5加密迭代两次
+            user.setPassword(UserCredentialsMatcher.encryptPassword("md5", "123456", salt, 2));
+            user.setSalt(salt);
+            user.setAvailable(user.getAvailable() == null ? 0 : 1);
+            switch (user.getType()) {
+                case 0:
+                    user.setPosition("系统管理员");
+                    break;
+                case 1:
+                    user.setPosition("数据录入员");
+                    break;
+                case 2:
+                    user.setPosition("数据分析员");
+                    break;
+                case 3:
+                    user.setPosition("普通用户");
+                    break;
+
+            }
+            int i = userService.addUser(user);
+            if (i > 0) {
+                jo.put("code", 0);
+                jo.put("message", "添加成功,初始密码123456");
+            } else {
+                jo.put("code", 2);
+                jo.put("message", "数据插入异常");
+            }
+            return jo;
+        }
+    }
+
 
     @RequestMapping(value = "/user/isNameExist", method = POST)
     @ResponseBody
@@ -117,7 +166,7 @@ public class UserController {
     @RequestMapping("logout")
     public String logout() {
         SecurityUtils.getSubject().logout();
-        return "view/login";
+        return "view/toLogin";
     }
 
     @RequestMapping("/user/manageUser")
@@ -133,7 +182,8 @@ public class UserController {
 
     @RequestMapping("/user/changePassword")
     @ResponseBody
-    public String changePassword(@RequestParam(required = true) int id, @RequestParam(required = true) String pwd, @RequestParam(required = true, defaultValue = "123456") String pwd1) {
+    public JSONObject changePassword(@RequestParam(required = true) int id, String pwd, String pwd1) {
+        JSONObject jo = new JSONObject();
         User currentLoginUser = (User) SecurityUtils.getSubject().getPrincipal();
         Subject subject = SecurityUtils.getSubject();
         User operateUser = null;
@@ -142,10 +192,22 @@ public class UserController {
         } else if (subject.isPermitted("user:updateProfile")) {
             operateUser = currentLoginUser;
         } else {
-            //没有权限修改
-            return "2";
+            jo.put("code", 2);
+            jo.put("message", "没有权限修改");
+            return jo;
         }
-        boolean isOldPwdRight = UserCredentialsMatcher.encryptPassword("md5", pwd, operateUser.getSalt(), 2).equals(operateUser.getPassword());
+
+        boolean isOldPwdRight = false;
+        if (pwd != null) {
+            isOldPwdRight = UserCredentialsMatcher.encryptPassword("md5", pwd, operateUser.getSalt(), 2).equals(operateUser.getPassword());
+        }
+        if (pwd == null || pwd1 == null) {
+            if (subject.isPermitted("user:manageUser")) {
+                isOldPwdRight = true;
+                pwd1 = "123456";
+            }
+        }
+
         if (isOldPwdRight) {
             //重新生成随机Salt
             String salt = UserCredentialsMatcher.generateSalt(6);
@@ -153,33 +215,37 @@ public class UserController {
             String newPassword = UserCredentialsMatcher.encryptPassword("md5", pwd1, salt, 2);
             int i = userService.changePassword(operateUser.getId(), newPassword, salt);
             if (i > 0) {
-                return "1";
+                jo.put("code", 1);
+                jo.put("message", "修改成功");
+                sessionService.stopSessionByUserid(id);
+                return jo;
             } else {
-                //异常导致更新密码失败
-                return "0";
+                jo.put("code", 0);
+                jo.put("message", "异常导致更新密码失败");
+                return jo;
             }
         } else {
-            //旧密码输入错误
-            return "3";
+            jo.put("code", 3);
+            jo.put("message", "旧密码输入错误");
+            return jo;
         }
-
     }
 
     @RequestMapping("/user/updateUser")
     @ResponseBody
-    public String updateUser(User user) {
+    public JSONObject updateUser(User user) {
         Subject subject = SecurityUtils.getSubject();
         User operateUser = (User) subject.getPrincipal();
-
-        if(subject.isPermitted("user:manageUser") && operateUser.getId() != user.getId()){
+        JSONObject jo = new JSONObject();
+        if (subject.isPermitted("user:manageUser") && operateUser.getId() != user.getId()) {
             operateUser = userService.findUserByUserId(user.getId());
             operateUser.setPosition(user.getPosition());
             operateUser.setType(user.getType());
-            operateUser.setAvailable(user.getAvailable());
+            operateUser.setAvailable(user.getAvailable() == null ? 0 : 1);
             operateUser.setUsername(user.getUsername());
         }
 
-        if(subject.isPermitted("user:updateProfile")){
+        if (subject.isPermitted("user:updateProfile")) {
             operateUser.setRealname(user.getRealname());
             operateUser.setSex(user.getSex());
             operateUser.setPhone(user.getPhone());
@@ -189,26 +255,61 @@ public class UserController {
         int i = userService.updateUser(operateUser);
         if (i > 0) {
             //修改当前登录用户的个人信息
-            if(operateUser.getId() == ((User)subject.getPrincipal()).getId()){
+            if (operateUser.getId() == ((User) subject.getPrincipal()).getId()) {
                 session.setAttribute("user", operateUser);
-            }else{
+            } else {
                 //管理员修改其他用户的登录信息
-                SessionService.stopSessionByUserid(operateUser.getId());
+                sessionService.stopSessionByUserid(operateUser.getId());
             }
-            return "修改成功";
+            jo.put("code", 0);
+            jo.put("message", "修改成功");
         } else {
-            return "修改失败";
+            jo.put("code", 1);
+            jo.put("message", "修改失败");
         }
+        return jo;
+    }
+
+    @RequestMapping("/user/selectAllUser")
+    @ResponseBody
+    public Map<String, Object> selectAllUser(@RequestParam(required = false) String username,@RequestParam(required = false) String realname, @RequestParam Integer page, @RequestParam Integer limit) {
+        User user = new User();
+        user.setUsername(username);
+        user.setRealname(realname);
+        List<User> users = userService.findAll(user);
+        List<User> subUsers = new ArrayList<>(limit);
+        for (int i = 0; i < limit && (page - 1) * limit + i < users.size(); i++) {
+            User u = users.get((page - 1) * limit + i);
+            if (u == null) break;
+            else subUsers.add(u);
+        }
+        HashMap<String, Object> map = new HashMap<>();
+        map.put("code", 0);
+        map.put("count", users.size());
+        map.put("msg", "");
+        map.put("data", subUsers);
+        return map;
     }
 
     @RequestMapping("/user/deleteUser")
     @ResponseBody
-    public String deleteUser(int id) {
+    //public String deleteUser(@RequestBody JSONObject jsonObject) {
+    public JSONObject deleteUser(int id) {
+        JSONObject jo = new JSONObject();
+        if (id == ((User) SecurityUtils.getSubject().getPrincipal()).getId()) {
+            jo.put("code", 2);
+            jo.put("message", "不能删除自己");
+            return jo;
+        }
         int i = userService.deleteUser(id);
         if (i > 0) {
-            return "删除成功";
+            jo.put("code", 0);
+            jo.put("message", "删除成功");
+            return jo;
         } else {
-            return "删除失败";
+            jo.put("code", 1);
+            jo.put("message", "删除错误");
+            return jo;
         }
     }
 }
